Refactory types
