+++ src/components/index.ts
@@ -1,2 +1,3 @@
 export * from "./nav-bar";
 export * from "./characters";
+export * from "./types";
